\name{loadRData}
\alias{loadRData}
\title{loadRData - Load all objects saved in the most recent archive}
\usage{
loadRData(dir, prefix = "R")
}
\arguments{
  \item{dir}{Directory in which the objects will be loaded}
}
\value{
Type NULL

none
}
\description{
loadRData - Load all objects saved in the most recent
archive
}
\author{
matheus
}

