\name{multiplot}
\alias{multiplot}
\title{Do multiple plots in one image
http://www.cookbook-r.com/}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
  \item{...}{plots objects from ggplot2 (gg object)}

  \item{plotlist}{a list of plots objects from ggplot2 (gg
  object)}

  \item{cols}{Number of columns in layout}

  \item{layout}{A matrix specifying the layout. If present,
  'cols' is ignored.}
}
\description{
Do multiple plots in one image http://www.cookbook-r.com/
}
\references{
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}

