\name{saveRData}
\alias{saveRData}
\title{saveRData - Save all objects in a archive, named with pattern name.}
\usage{
saveRData(dir, prefix = "R")
}
\arguments{
  \item{dir}{Directory in which the objects will be saved}
}
\value{
Type NULL

none
}
\description{
saveRData - Save all objects in a archive, named with
pattern name.
}
\author{
matheus, daniel
}

